+++ app/src/main/AndroidManifest.xml
@@ -13,15 +13,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.reservation.ReservationActivity"
+            android:name=".view.reservation.detail.ReservationActivity"
             android:exported="false"
-            android:parentActivityName=".view.MainActivity"/>
+            android:parentActivityName=".view.movies.MainActivity"/>
         <activity
-            android:name=".view.reservation.ReservationCompleteActivity"
+            android:name=".view.reservation.result.ReservationCompleteActivity"
             android:exported="false"
-            android:parentActivityName=".view.MainActivity" />
+            android:parentActivityName=".view.movies.MainActivity" />
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.movies.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
