+++ app/src/main/java/woowacourse/movie/view/movies/OnMovieEventListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.movies
 
 import woowacourse.movie.domain.Movie
