+++ app/src/main/res/layout/activity_reservation_complete.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.reservation.ReservationCompleteActivity">
+    tools:context=".view.reservation.result.ReservationCompleteActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
